[FX] refactor experimental tracer and adapt it with hf models
