+++ colossalai/_analyzer/fx/__init__.py
@@ -1,5 +1,3 @@
-from .bias_addition import *
-from .custom_leaf_module import *
 from .node_util import MetaInfo
 from .symbolic_profile import symbolic_profile
-from .symbolic_trace import symbolic_trace
+from .tracer.symbolic_trace import symbolic_trace
