+++ colossalai/_analyzer/fx/tracer/bias_addition.py
@@ -7,7 +7,7 @@ import torch
 import torch.nn.functional as F
 from torch.nn.modules.utils import _pair, _single, _triple
 
-from .symbolic_trace import register_tracer_impl
+from .tracer import register_tracer_impl
 
 __all__ = []
