+++ colossalai/_analyzer/fx/tracer/tracer.py
@@ -1,28 +1,18 @@
 import functools
 import inspect
-import operator
 from contextlib import contextmanager
-from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Tuple, Type, Union
+from typing import Any, Callable, Dict, Iterable, Optional, Set, Tuple, Type, Union
 
 import torch
 import torch.nn as nn
 from torch.fx import Graph, Node, Proxy, Tracer
-from torch.fx.graph import _Namespace
 from torch.utils._pytree import tree_map
 
-from colossalai._analyzer._subclasses import MetaTensor, _TensorPropertyMethod, _TorchFactoryMethod
+from colossalai._analyzer._subclasses import _TensorPropertyMethod, _TorchFactoryMethod
 
-from .codegen import ActivationCheckpointCodeGen
-from .graph_module import ColoGraphModule
-from .node_util import MetaInfo
+from .proxy import ColoProxy
 
 Target = Union[Callable[..., Any], str]
-Argument = Optional[Union[Tuple[Any, ...],    # actually Argument, but mypy can't represent recursive types
-                          List[Any],    # actually Argument
-                          Dict[str, Any],    # actually Argument
-                          slice,    # Slice[Argument, Argument, Argument], but slice is not a templated type in typing
-                          'Node',]]
-zeros = torch.zeros
 
 
 def _truncate_suffix(s: str):
@@ -30,17 +20,6 @@ def _truncate_suffix(s: str):
 
     # FIXME: don't know why but torch.fx always gets a suffix like '_1' in the name
     return re.sub(r'_\d+$', '', s)
-
-
-def _default_device():
-    return torch.device('cuda:0') if torch.cuda.is_available() else torch.device('cpu')
-
-
-def _current_device(module):
-    try:
-        return next(module.parameters()).device
-    except:
-        return _default_device()
 
 
 def register_tracer_impl(func: Callable[..., Any], name: Optional[str] = '_custom_impl'):
@@ -68,149 +47,6 @@ def register_leaf_module(module: nn.Module):
 
 def register_non_leaf_module(module: nn.Module):
     ColoTracer._custom_non_leaf_module.add(module)
-
-
-class ColoProxy(Proxy):
-    _func_dispatch: Dict[Target, Callable[..., Any]] = {}
-
-    def __init__(self, *args, data=None, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._meta_data = data
-
-    @property
-    def meta_data(self):
-        return self._meta_data
-
-    @meta_data.setter
-    def meta_data(self, args):
-        wrap_fn = lambda x: MetaTensor(x) if isinstance(x, torch.Tensor) else x
-        self._meta_data = tree_map(wrap_fn, args)
-
-    # @classmethod
-    # def __torch_function__(cls, orig_method, types, args=(), kwargs=None):
-    #     kwargs = {} if kwargs is None else kwargs
-    #     if orig_method in cls._func_dispatch:
-    #         impl = cls._func_dispatch.pop(orig_method)    # avoid recursion
-    #         proxy = impl(*args, **kwargs)
-    #         cls._func_dispatch[orig_method] = impl
-    #         return proxy
-    #     else:
-    #         proxy = cls.from_torch_proxy(super().__torch_function__(orig_method, types, args, kwargs))
-    #         unwrap_fn = lambda p: p.meta_data if isinstance(p, ColoProxy) else p
-    #         if proxy.meta_data is None:
-    #             proxy.meta_data = orig_method(*tree_map(unwrap_fn, args), **tree_map(unwrap_fn, kwargs))
-    #         return proxy
-
-    @classmethod
-    def from_torch_proxy(cls, proxy: Proxy):
-        return cls(proxy.node, proxy.tracer)
-
-    def __repr__(self):
-        return f"ColoProxy({self.node.name}, meta_data={self.meta_data})"
-
-    def __len__(self):
-        return len(self.meta_data)
-
-    def __int__(self):
-        return int(self.meta_data)
-
-    def __index__(self):
-        try:
-            return int(self.meta_data)
-        except:
-            return zeros(self.meta_data.shape, dtype=torch.bool).numpy().__index__()
-
-    def __float__(self):
-        return float(self.meta_data)
-
-    def __bool__(self):
-        return self.meta_data
-
-    def __getattr__(self, k):
-        return ColoAttribute(self, k, getattr(self._meta_data, k, None))
-
-    def __setitem__(self, key, value):
-        proxy = self.tracer.create_proxy('call_function', operator.setitem, (self, key, value), {})
-        proxy.meta_data = self._meta_data
-        return proxy
-
-    def __contains__(self, key):
-        if self.node.op == "placeholder":
-            # this is used to handle like
-            # if x in kwargs
-            # we don't handle this case for now
-            return False
-        return super().__contains__(key)
-
-    def __isinstancecheck__(self, type):
-        return isinstance(self.meta_data, type)
-
-    # def size(self, dim=None):
-    #     if self._meta_data is None:
-    #         return self._meta_data.size(*[dim] if dim else [])
-    #     return self.tracer.create_proxy('call_method', 'size', (self, dim) if dim else (self,), {})
-
-    # def dim(self):
-    #     if self._meta_data is not None:
-    #         return self._meta_data.dim()
-    #     return self.tracer.create_proxy('call_method', 'dim', (self,), {})
-
-    # @property
-    # def shape(self):
-    #     if self._meta_data is not None:
-    #         return self._meta_data.shape
-    #     return self.tracer.create_proxy('call_function', getattr, (self, 'shape'), {})
-
-    # @property
-    # def ndim(self):
-    #     if self._meta_data is not None:
-    #         return self._meta_data.ndim
-    #     return self.tracer.create_proxy('call_function', getattr, (self, 'ndim'), {})
-
-    # @property
-    # def device(self):
-    #     if self._meta_data is not None:
-    #         return self._meta_data.device
-    #     return self.tracer.create_proxy('call_function', getattr, (self, 'device'), {})
-
-    # @property
-    # def dtype(self):
-    #     if self._meta_data is not None:
-    #         return self._meta_data.dtype
-    #     return self.tracer.create_proxy('call_function', getattr, (self, 'dtype'), {})
-
-    # def to(self, *args, **kwargs):
-    #     return self.tracer.create_proxy('call_method', 'to', (self, *args), {**kwargs})
-
-    # def cpu(self, *args, **kwargs):
-    #     return self.tracer.create_proxy('call_method', 'cpu', (self, *args), {**kwargs})
-
-    # def cuda(self, *args, **kwargs):
-    #     return self.tracer.create_proxy('call_method', 'cuda', (self, *args), {**kwargs})
-
-
-class ColoAttribute(ColoProxy):
-
-    def __init__(self, root, attr: str, data=None):
-        self.root = root
-        self.attr = attr
-        self.tracer = root.tracer
-        self._meta_data = data
-        self._node: Optional[Node] = None
-
-    @property
-    def node(self):
-        # the node for attributes is added lazily, since most will just be method calls
-        # which do not rely on the getitem call
-        if self._node is None:
-            self._node = self.tracer.create_proxy('call_function', getattr, (self.root, self.attr), {}).node
-        return self._node
-
-    def __call__(self, *args, **kwargs):
-        return self.tracer.create_proxy('call_method', self.attr, (self.root,) + args, kwargs)
-
-    def __repr__(self):
-        return f"ColoAttribute({self.node.name}, attr={self.attr})"
 
 
 class ColoTracer(Tracer):
@@ -318,7 +154,7 @@ class ColoTracer(Tracer):
 
     def create_node(self, *args, **kwargs) -> Node:
         node = super().create_node(*args, **kwargs)
-        n_info = MetaInfo(node, mod_dir=self.mod_dir, to_recompute=tuple(self.ckpt_regions))
+        # n_info = MetaInfo(node, mod_dir=self.mod_dir, to_recompute=tuple(self.ckpt_regions))
         return node
 
     def trace(self,
@@ -524,134 +360,3 @@ class ColoTracer(Tracer):
                 return maybe_parameter_proxy
 
         return attr_val
-
-
-def symbolic_trace(
-    root: Union[torch.nn.Module, Callable[..., Any]],
-    concrete_args: Optional[Dict[str, Any]] = None,
-    meta_args: Optional[Dict[str, Any]] = None,
-    trace_act_ckpt: bool = False,
-    bias_addition_split: bool = False,
-) -> ColoGraphModule:
-    """
-    Traces a ``torch.nn.Module`` or a function and returns a ``GraphModule`` with ``Node``s and ``MetaInfo``
-    attached to the ``Node``s.
-
-    Can be used to trace the usage of ``torch.utils.checkpoint`` and the path of module
-    (https://github.com/pytorch/examples/blob/main/fx/module_tracer.py).
-
-    This tracer is able to trace basic control flow and for loops.
-
-    It will split the bias addition into two parts if ``bias_addition_split`` is set to be ``True``.
-    (See ./bias_addition.py for more details).
-
-    Examples:
-    1. Tracing a ``torch.nn.Module`` with control flow.
-
-    .. code-block:: python
-
-        class MyModule(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = torch.nn.Linear(2, 2)
-
-            def forward(self, x):
-                if x.size(0) > 1:
-                    x = x.sum(dim=0)
-                return self.linear(x)
-
-        traced = symbolic_trace(MyModule(), meta_args={'x': torch.randn(1, 2, 2)})
-
-        # traced code like:
-        # def forward(self, x):
-        #     linear_1 = self.linear(x)
-        #     return linear_1
-
-        traced = symbolic_trace(MyModule(), meta_args={'x': torch.randn(2, 2, 2)})
-
-        # traced code like:
-        # def forward(self, x):
-        #     sum = x.sum(dim=0); x = None
-        #     linear = self.linear(sum); sum = None
-        #     return linear
-
-    2. Tracing a ``torch.nn.Module`` with ``torch.utils.checkpoint``.
-
-    .. code-block:: python
-
-        class MyModule(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = torch.nn.Linear(2, 2)
-
-            def forward(self, x):
-                def custom_forward(x):
-                    return self.linear(x)
-                return torch.utils.checkpoint.checkpoint(custom_forward, x)
-
-        traced = symbolic_trace(MyModule(), meta_args={'x': torch.randn(1, 2, 2)}, trace_act_ckpt=True)
-
-        # traced code like:
-        # def checkpoint_0(self, x):
-        #     linear = self.linear(x); x = None
-        #     return linear
-        #
-        # def forward(self, x):
-        #     linear = torch.utils.checkpoint.checkpoint(checkpoint_0, x); x = None
-        #     return linear
-
-    3. Tracing a ``torch.nn.Module`` with ``bias_addition_split``.
-
-    .. code-block:: python
-
-        class MyModule(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = torch.nn.Linear(2, 2, bias=True)
-
-            def forward(self, x):
-                return self.linear(x)
-
-        traced = symbolic_trace(MyModule(), meta_args={'x': torch.randn(1, 2, 2)}, bias_addition_split=True)
-
-        # traced code like:
-        # def forward(self, x):
-        #     linear_bias = self.linear.bias
-        #     linear_weight = self.linear.weight
-        #     linear = torch._C._nn.linear(x, linear_weight);  x = linear_weight = None
-        #     add = linear + linear_bias;  linear = linear_bias = None
-        #     return add
-
-    Args:
-        root (Union[torch.nn.Module, Callable[..., Any]]): The ``torch.nn.Module`` or function to be traced.
-        concrete_args (Optional[Dict[str, Any]], optional): Concrete arguments to be passed to the ``root``.
-            Defaults to {}.
-        meta_args (Optional[Dict[str, Any]], optional): Meta arguments to be passed to the ``root``. Mostly used
-            for tracing control flow. Defaults to {}.
-        trace_act_ckpt (bool, optional): Whether to trace the usage of ``torch.utils.checkpoint``.
-            Defaults to False.
-        bias_addition_split (bool, optional): Whether to split the bias addition into two parts. Defaults to False.
-
-    Returns:
-        ColoGraphModule: A traced ``GraphModule`` that is ready for activation checkpoint ``CodeGen``.
-
-    Remarks:
-        This part of ``symbolic_trace()`` is maintained by Colossal-AI team. If you encountered
-        any unexpected error during tracing, feel free to raise an issue on Colossal-AI GitHub
-        repo. We welcome any feedback and contributions to enhance the extensibility of
-        Colossal-AI.
-    """
-    if meta_args:
-        device, orig_device = _default_device(), _current_device(root)
-        wrap_fn = lambda elem: MetaTensor(elem, device=device) if isinstance(elem, torch.Tensor) else elem
-        graph = ColoTracer(trace_act_ckpt=trace_act_ckpt,
-                           bias_addition_split=bias_addition_split).trace(root.to(device),
-                                                                          concrete_args=concrete_args,
-                                                                          meta_args=tree_map(wrap_fn, meta_args))
-        if trace_act_ckpt:
-            graph.set_codegen(ActivationCheckpointCodeGen())
-        root.to(orig_device)
-    else:
-        graph = Tracer().trace(root, concrete_args=concrete_args)
-    name = root.__class__.__name__ if isinstance(root, torch.nn.Module) else root.__name__
-    return ColoGraphModule(root, graph, name)
